Consolidate persistence and storage helpers under Core/Persistence
Rename Persistence to PersistentContainer, fold in its load and background-task extensions, and move the keychain and registry storage helpers out of Core/Utilities/Storage into Core/Persistence. Mirror the move in the test tree.
