+++ Tests/ScoutTests/Core/Persistence/RegistryTests.swift
